xtensa: fix register access using PTRACE_GETREGS API
The commit 2429c69961e2 ("xtensa: switch to PTRACE_GETREGS API") changed
the way xtensa registers are accessed, but the change is only correct
in case when the xtensa_regs.windowbase is 0 (i.e. the current register
window starts at the physical registr 0). In other cases decoded syscall
arguments are bogus. This issue is currently observed on noMMU xtensa
linux because it uses PTRACE_GETREGS API, while configurations with MMU
default to using ptrace_get_syscall_info.

Fix that by taking xtensa_registers.windowbase into an account.
This is trivial for the first 4 registers in the current window, because
windowbase indicates 4-register groups and is always valid, so registers
a0...a3 are guaranteed to be available in the physical registers array
at indices windowbase * 4 + 0...3.
Position of the registers with higher numbers may wrap around the end of
the physical register file, so e.g. with 32 physical resgisters and
windowbase = 7 the register a5 is stored in the xtensa_registers.a[1].

This is further complicated by the fact that the size of the physical
register file may vary and there's no explicit way to inquire it from
the kernel. Probe for it by trying to set windowbase to ascending powers
of 2 until it fails.

* src/linux/xtensa/arch_regs.c (ARCH_SP_REG): Account for
xtensa_regs.windowbase in xtensa_regs.a indexing.
* src/linux/xtensa/get_error.c (arch_get_error): New variable
status_reg that accounts for xtensa_regs.windowbase, use it
for xtensa_regs.a indexing.
* src/linux/xtensa/get_syscall_args.c (xtensa_probe_naregs): New
function.
(arch_get_syscall_args): New variable naregs_mask, initialize it
using xtensa_probe_naregs, use it for xtensa_regs.a indexing.
* src/linux/xtensa/set_error.c (arch_set_error)
(arch_set_success): Account for xtensa_regs.windowbase in
xtensa_regs.a indexing.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>