tests/init.sh: add --status to pidns strace call
The pidns test harness processes multi-process log output,
which is inherently prone to ordering issues like this:

    -7     get_robust_list(2 /* 7 in strace's PID NS */, [0x7ff3fc9a7a20], [24]) = 0
    +7     get_robust_list(2 /* 7 in strace's PID NS */,  <unfinished ...>
    +7     <... get_robust_list resumed>[0x7ff3fc9a7a20], [24]) = 0
    FAIL xet_robust_list--pidns-translation.gen.test (exit status: 1)

Try to avoid them by forcing the buffering of the syscall trace output
with the --status option that excludes the "unavailable" status,
that is not expected in the test output.

* tests/init.sh (test_pidns_run_strace): Add "--status=!unavailable"
option to the strace call.